resolved warnings, which arose during python setup.py build

+++ src/core/primitives.cc
@@ -824,7 +824,7 @@ std::unique_ptr<const Geometry> PolygonNode::createGeometry() const
   if (this->paths.empty() && this->points.size() > 2) {
     Outline2d outline;
     std::vector<size_t> path;
-    for (int i = 0; i < this->points.size(); i++) path.push_back(i);
+    for (size_t i = 0; i < this->points.size(); i++) path.push_back(i);
     outline.vertices = createGeometry_sub(this->points, path, fn, fa, fs);
     p->addOutline(outline);
   } else {
